requirements.txt
# ── Phase 1: Preprocessing + Training ──
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
python-dotenv==1.1.0

# --- Training & Experiment Tracking
imbalanced-learn==0.13.0
mlflow==2.21.3

# ── Phase 2: Serving ──
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
python-multipart==0.0.20

# ── Phase 3: Testing ──
pytest==8.3.5
httpx==0.28.1